+++ src/slib/java/org/smoothbuild/builtin/file/match/Constants.java
@@ -1,7 +1,0 @@
-package org.smoothbuild.builtin.file.match;
-
-public class Constants {
-  public static final char SINGLE_STAR_CHAR = '*';
-  public static final String SINGLE_STAR = "*";
-  public static final String DOUBLE_STAR = "**";
-}