+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change Log
 Version 0.19.0 (????.??.??)
 ---------------------------
 
+ * improved `filter` function to support more glob wildcards
  * renamed 'concatenate' function to 'concat'
  * added parallel execution
  * removed 'javaProperty' function from standard library (aka platform API)